Refactoring: removal of commands/helpers
we do not need it basically. The entire logic belongs to the generate command.

Constructor of Generator belongs to the generator package. (We use the external testing approach, to test only the exposed APIs)
